fixing the tests
